Split SqlServerDrillDownCollector + Lite DrillDownCollector into per-domain partials
Behavior-neutral partial-class split, done for BOTH apps for parity.

Each non-partial collector becomes `partial`, with its instance Collect* drill-down
methods fanned (verbatim, no reformatting) into 5 per-domain sibling files:
Blocking / Queries / Config / Storage / Plans. Everything else stays in the base
(ctor, EnrichFindingsAsync, the static CollectPlanNodes / CollectRcsiInactionFigures
helpers, and the nested ConfigIssueRow / RcsiInactionFigures types). No call-site
changes; same class, same namespace.

- Dashboard: 19 methods moved across the 5 partials.
- Lite: 17 methods moved across the 5 partials.

One source-introspection test (RemediationTests.Detector_CpuPercent_IsRealWindowShare_
NotHardcodedZero) read the base collector file for SQL that now lives in the Queries
partial; updated it to concatenate all SqlServerDrillDownCollector*.cs partials so its
intent is preserved regardless of which partial holds CollectAbnormalCpuPlans.

Verified: solution builds (0 warnings, 0 errors); Dashboard.Tests 491 passed / 0 failed;
Lite.Tests 545 passed / 0 failed.
